Move BrowserPluginGuestDelegate to a separate class
