+++ filenames.gypi
@@ -207,6 +207,8 @@
       'atom/browser/ui/x/window_state_watcher.h',
       'atom/browser/ui/x/x_window_utils.cc',
       'atom/browser/ui/x/x_window_utils.h',
+      'atom/browser/web_view_guest_delegate.cc',
+      'atom/browser/web_view_guest_delegate.h',
       'atom/browser/web_view_manager.cc',
       'atom/browser/web_view_manager.h',
       'atom/browser/web_dialog_helper.cc',
